changed id from int to Long

+++ Kamienica/src/main/java/kamienica/dao/DaoInterface.java
@@ -9,11 +9,11 @@ public interface DaoInterface<T extends Object> {
 
 	public List<T> getList();
 
-	public void deleteById(int id);
+	public void deleteById(Long id);
 
 	public void update(T object);
 
-	public T getById(int id);
+	public T getById(Long id);
 	
 
 }
