+++ Kamienica/src/main/java/kamienica/feature/division/DivisionServiceImpl.java
@@ -18,7 +18,7 @@ public class DivisionServiceImpl implements DivisionService {
 	}
 
 	@Override
-	public void deleteByID(int id) {
+	public void deleteByID(Long id) {
 		divisionDAO.deleteById(id);
 
 	}
