+++ Kamienica/src/main/java/kamienica/feature/reading/ReadingWaterConverter.java
@@ -19,7 +19,7 @@ public class ReadingWaterConverter implements Converter<Object, ReadingWater> {
 
 	@Override
 	public ReadingWater convert(Object element) {
-		Integer id = Integer.parseInt((String) element);
+		Long id = Long.parseLong((String) element);
 		return readingService.getWaterById(id);
 	}
